requirements.txt
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
cryptography==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.5
django-allauth==0.44.0
django-appconf==1.0.4
django-ckeditor==6.1.0
django-crispy-forms==1.10.0
django-eav2==0.13.0
django-imagekit==4.0.2
django-import-export==2.5.0
django-js-asset==1.2.2
et-xmlfile==1.1.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
pilkit==2.0
Pillow==8.1.0
pycparser==2.20
PyJWT==2.0.1
pylint==2.6.0
python3-openid==3.2.0
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
stripe==2.56.0
tablib==3.0.0
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
